+++ src/tiled/map/spawn.rs
@@ -4,9 +4,13 @@
 //! It handles the creation of map layers, tiles, objects, and their associated components in the ECS world,
 //! enabling the rendering and interaction of Tiled maps within a Bevy application.
 
+#[cfg(feature = "text")]
+use crate::tiled::object::{TiledObjectText, TiledObjectTextBox};
 use crate::{prelude::*, tiled::event::TiledMessageWriters};
-use bevy::{platform::collections::HashMap, prelude::*, sprite::Anchor};
+#[cfg(feature = "text")]
+use bevy::camera::primitives::Aabb;
 
+use bevy::{platform::collections::HashMap, prelude::*, sprite::Anchor};
 #[cfg(feature = "render")]
 use bevy_ecs_tilemap::prelude::{TilemapBundle, TilemapSpacing};
 
@@ -408,7 +412,7 @@ fn spawn_objects_layer(
         let object_kind = match tiled_object {
             TiledObject::Point => "Point",
             TiledObject::Tile { .. } => "Tile",
-            TiledObject::Text => "Text",
+            TiledObject::Text { .. } => "Text",
             TiledObject::Rectangle { .. } => "Rectangle",
             TiledObject::Ellipse { .. } => "Ellipse",
             TiledObject::Polygon { .. } => "Polygon",
@@ -421,7 +425,7 @@ fn spawn_objects_layer(
                 ChildOf(layer_event.origin),
                 TiledMapReference(layer_event.get_map_entity().unwrap()),
                 TiledObjectId(object_data.id()),
-                tiled_object,
+                tiled_object.clone(),
                 TiledName(object_data.name.clone()),
                 transform,
                 match &object_data.visible {
@@ -430,6 +434,42 @@ fn spawn_objects_layer(
                 },
             ))
             .id();
+
+        #[cfg(feature = "text")]
+        if let TiledObject::Text {
+            width,
+            height,
+            text,
+            offset,
+            font_size,
+        } = tiled_object
+        {
+            commands.spawn((
+                ChildOf(object_entity),
+                Name::new("TiledObjectTextBox"),
+                TiledObjectTextBox,
+                Aabb::from_min_max(Vec3::ZERO, Vec3::new(width, height, 0.)),
+                Transform::from_translation(Vec3::ZERO),
+                Anchor::TOP_LEFT,
+                Visibility::Inherited,
+                children![(
+                    Name::new("TiledObjectText"),
+                    TiledObjectText,
+                    Text2d::new(text),
+                    TextFont {
+                        font_size: font_size as f32,
+                        ..default()
+                    },
+                    Anchor::from(offset),
+                    Transform::from_translation(Vec3::new(
+                        offset.x * width + width / 2.,
+                        offset.y * height - height / 2.,
+                        1.
+                    )),
+                    Visibility::Inherited,
+                )],
+            ));
+        }
 
         // Handle objects containing tile data:
         // we want to add a Sprite component to the object entity