+++ src/tiled/object.rs
@@ -4,6 +4,7 @@
 
 use crate::prelude::{geo::Centroid, *};
 use crate::tiled::helpers::iso_projection;
+use ::tiled::{HorizontalAlignment, VerticalAlignment};
 use bevy::prelude::*;
 
 /// Relationship and Marker [`Component`] for the visual representation of a [`TiledObject`].
@@ -28,6 +29,14 @@ pub struct TiledObjectVisuals(Vec<Entity>);
 #[derive(Component, Default, Reflect, Copy, Clone, Debug, Deref)]
 #[reflect(Component, Default, Debug)]
 pub struct TiledObjectId(pub u32);
+
+/// Marker [`Component`] for a [`TiledObject`]'s text box.
+#[derive(Component, Reflect, Copy, Clone, Debug)]
+pub struct TiledObjectTextBox;
+
+/// Marker [`Component`] for a [`TiledObjectTextBox`]'s inner text.
+#[derive(Component, Reflect, Copy, Clone, Debug)]
+pub struct TiledObjectText;
 
 /// Marker [`Component`] for a Tiled map object.
 #[derive(Component, Default, Reflect, Clone, Debug)]
@@ -77,9 +86,18 @@ pub enum TiledObject {
         height: f32,
     },
     /// A text object, which contains text data.
-    ///
-    /// Not supported yet.
-    Text,
+    Text {
+        /// the width of the text box.
+        width: f32,
+        /// the height of the text box.
+        height: f32,
+        /// the text content.
+        text: String,
+        /// the vertical and horizontal alignment of the text.
+        offset: Vec2,
+        /// the font size in pixels.
+        font_size: usize,
+    },
 }
 
 impl TiledObject {
@@ -112,9 +130,51 @@ impl TiledObject {
                 tiled::ObjectShape::Polyline { points } => TiledObject::Polyline {
                     vertices: points.into_iter().map(|(x, y)| Vec2::new(x, -y)).collect(),
                 },
-                tiled::ObjectShape::Text { .. } => {
-                    log::warn!("Text objects are not supported yet");
-                    TiledObject::Text
+                tiled::ObjectShape::Text {
+                    width,
+                    height,
+                    text,
+                    valign,
+                    halign,
+                    pixel_size,
+                    ..
+                } => {
+                    let offset = match (valign, halign) {
+                        (VerticalAlignment::Top, HorizontalAlignment::Left)
+                        | (VerticalAlignment::Top, HorizontalAlignment::Justify) => {
+                            Vec2::new(-0.5, 0.5)
+                        }
+                        (VerticalAlignment::Bottom, HorizontalAlignment::Left)
+                        | (VerticalAlignment::Bottom, HorizontalAlignment::Justify) => {
+                            Vec2::new(-0.5, -0.5)
+                        }
+                        (VerticalAlignment::Center, HorizontalAlignment::Left)
+                        | (VerticalAlignment::Center, HorizontalAlignment::Justify) => {
+                            Vec2::new(-0.5, 0.0)
+                        }
+                        (VerticalAlignment::Top, HorizontalAlignment::Right) => Vec2::new(0.5, 0.5),
+                        (VerticalAlignment::Bottom, HorizontalAlignment::Right) => {
+                            Vec2::new(0.5, -0.5)
+                        }
+                        (VerticalAlignment::Center, HorizontalAlignment::Right) => {
+                            Vec2::new(0.5, 0.0)
+                        }
+                        (VerticalAlignment::Top, HorizontalAlignment::Center) => {
+                            Vec2::new(0.0, 0.5)
+                        }
+                        (VerticalAlignment::Bottom, HorizontalAlignment::Center) => {
+                            Vec2::new(0.0, -0.5)
+                        }
+                        (VerticalAlignment::Center, HorizontalAlignment::Center) => Vec2::ZERO,
+                    };
+
+                    TiledObject::Text {
+                        width,
+                        height,
+                        text,
+                        offset,
+                        font_size: pixel_size,
+                    }
                 }
             }
         }
@@ -199,7 +259,7 @@ impl TiledObject {
 
         // Generate shape vertices relative to origin
         match self {
-            TiledObject::Point | TiledObject::Text => vec![Vec2::ZERO],
+            TiledObject::Point => vec![Vec2::ZERO],
             TiledObject::Tile { width, height } => {
                 vec![
                     Vec2::new(0., 0.),          // Bottom-left relative to object
@@ -208,7 +268,7 @@ impl TiledObject {
                     Vec2::new(*width, 0.),      // Bottom-right
                 ]
             }
-            TiledObject::Rectangle { width, height } => {
+            TiledObject::Rectangle { width, height } | TiledObject::Text { width, height, .. } => {
                 vec![
                     Vec2::new(0., 0.),           // Top-left relative to object
                     Vec2::new(*width, 0.),       // Top-right
@@ -288,9 +348,10 @@ impl TiledObject {
             offset,
         );
         match self {
-            TiledObject::Point | TiledObject::Text => None,
+            TiledObject::Point => None,
             TiledObject::Ellipse { .. }
             | TiledObject::Rectangle { .. }
+            | TiledObject::Text { .. }
             | TiledObject::Tile { .. }
             | TiledObject::Polygon { .. } => {
                 let mut line_string = geo::LineString::from(coords);